requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pydantic>=1.10.0

# Core ML / retrieval
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.6.0
faiss-cpu>=1.7.4
numpy>=1.24.0

# Utilities
rapidfuzz>=2.13.0
requests>=2.28.0
langdetect>=1.0.9
Unidecode>=1.3.6

# Optional / dev
tqdm>=4.65.0
